Move operations to op.raster.extent package.
